Add EmoteWhitelistHashManager and Emote moderation functionality
Allows you to turn a server side version of Emotecraft into "whitelist mode", this makes it so only emotes that the server admin uploads to the server-side whitelist directory can be used by players on the server. Emotes that aren't whitelisted will be canceled and won't be displayed to other users.

This is accomplished efficiently by checking the hash of the emote, which is generated by checking all of the emotes unique properties, namely all boneAnimations frames, any sounds, any particles, and the emote loop type.

How these properties are structured could easily change version to version of Emotecraft, so it's only guaranteed that this will work if the client and server are on the same version of Emotecraft.

Server side hashes are only taken once, unless an emote is modified, hashes are refreshed when the server is started, or via /emotes whitelist reload. You can also force a new hash to be taken of every emote via /emotes whitelist force-reload.